+++ test/PageContextSpec.js
@@ -9,9 +9,10 @@
 
 import React, { PropTypes } from 'react';
 import ReactDOM from 'react-dom';
+import ReactDOMServer from 'react-dom/server';
 import jsdom from 'mocha-jsdom';
 import { expect } from 'chai';
-import PageContext from '../src';
+import { PageContext } from '../src';
 
 describe('PageContext', () => {
 
@@ -24,12 +25,12 @@ describe('PageContext', () => {
     const container = document.getElementById('root');
     const Component = (props, { page }) => {
       page({ title: 'test 1' });
-      return <div />;
+      return <div>test</div>;
     };
     Component.contextTypes = { page: PropTypes.func.isRequired };
 
     ReactDOM.render(<PageContext><Component /></PageContext>, container, () => {
-      expect(container.innerHTML).to.be.equal('<div data-reactroot=""></div>');
+      expect(container.innerHTML).to.be.equal('<div data-reactroot="">test</div>');
       expect(document.title).to.be.equal('test 1');
       done();
     });
@@ -52,28 +53,87 @@ describe('PageContext', () => {
     });
   });
 
-  it('should read page metadata via ref={c => c.page}', (done) => {
-    let pageCtx;
+  it('should set description', (done) => {
     const container = document.getElementById('root');
     const Component = (props, { page }) => {
-      page({ title: 'test 2' });
+      page({ description: 'desc 2' });
+      return <div />;
+    };
+    Component.contextTypes = { page: PropTypes.func.isRequired };
+
+    ReactDOM.render(<PageContext><Component /></PageContext>, container, () => {
+      expect(document.title).to.be.empty;
+      expect(document.head.innerHTML)
+        .to.be.equal('<meta name="description" content="desc 2">');
+      done();
+    });
+  });
+
+  it('should set other meta tags', (done) => {
+    const container = document.getElementById('root');
+    const Component = (props, { page }) => {
+      page({
+        meta: [
+          { property: '1', content: '2' },
+          { property: '3', content: '4' },
+          { httpEquiv: '5', content: '6' },
+        ],
+      });
+      return <div />;
+    };
+    Component.contextTypes = { page: PropTypes.func.isRequired };
+
+    ReactDOM.render(<PageContext><Component /></PageContext>, container, () => {
+      expect(document.head.innerHTML)
+        .to.be.equal(
+          '<meta property="1" content="2">' +
+          '<meta property="3" content="4">' +
+          '<meta http-equiv="5" content="6">'
+        );
+      done();
+    });
+  });
+
+  it('should read page metadata via ref={c => c.page}', (done) => {
+    let result;
+    const container = document.getElementById('root');
+    const Component = (props, { page }) => {
+      page({
+        title: 'test 3',
+        meta: [{ name: 'description', content: 'desc 3' }],
+      });
       return <div />;
     };
 
     Component.contextTypes = { page: PropTypes.func.isRequired };
 
     ReactDOM.render(
-      <PageContext ref={c => { pageCtx = c.page; }}>
+      <PageContext ref={c => { result = c.page(); }}>
         <Component />
       </PageContext>,
       container,
       () => {
-        expect(document.title).to.be.equal('test 2');
-        expect(pageCtx).to.be.ok.and.a('function');
-        expect(pageCtx()).to.be.deep.equal({ title: 'test 2' });
+        expect(document.title).to.be.equal('test 3');
+        expect(result).to.be.an('object');
+        expect(result).to.have.property('title', 'test 3');
+        expect(result.meta).to.be.deep.equal([{ name: 'description', content: 'desc 3' }]);
         done();
       });
   });
 
-
+  it('should work with ReactDOM.renderToString()', () => {
+    let result;
+    const Component = (props, { page }) => {
+      page({ title: 'test 4' });
+      return <div />;
+    };
+    Component.contextTypes = { page: PropTypes.func.isRequired };
+    ReactDOMServer.renderToString(
+      <PageContext onChange={x => (result = x)}>
+        <Component />
+      </PageContext>
+    );
+    expect(result).to.be.an('object');
+    expect(result).to.have.property('title', 'test 4');
+  });
 });